+++ src/VDF/Champs/Champ_Don_Face_lu.cpp
@@ -101,7 +101,7 @@ Entree& Champ_Don_Face_lu::readOn(Entree& is)
         }
     }
 
-  int erreur = mp_sum(nb_faces - nb_faces_affectees);
+  int erreur = static_cast<int>(mp_sum(nb_faces - nb_faces_affectees)); // should remain small
   if (erreur)
     {
       Cerr << "Erreur dans Champ_Don_Face_lu.\n" << erreur << " faces n'ont pas ete affectees.\n" << "Voir la liste des faces dans les fichiers .log" << finl;
